remove test file and move the tests to dedicated test file
